Refactor paykit-lib modules and bump rc12
