+++ paykit-ffi/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "paykit-ffi"
-version = "0.1.0-rc11"
+version = "0.1.0-rc12"
 edition = "2021"
 rust-version = "1.91.1"
 description = "UniFFI bindings for paykit-lib targeting iOS and Android"
