+++ paykit-ffi/bindings/android/gradle.properties
@@ -3,4 +3,4 @@ android.useAndroidX=true
 android.enableJetifier=true
 kotlin.code.style=official
 group=com.synonym
-version=0.1.0-rc11
+version=0.1.0-rc12
