+++ paykit-lib/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "paykit-lib"
-version = "0.1.0-rc11"
+version = "0.1.0-rc12"
 authors = ["denys@synonym.to"]
 description = "Paykit library implementation"
 license = "MIT"
