+++ paykit-lib/src/encrypted_link.rs
@@ -1,0 +1,15 @@
+mod handshake;
+mod link;
+mod paths;
+mod snapshot;
+
+pub use handshake::{
+    accept_encrypted_link, advance_handshake, initiate_encrypted_link,
+    restore_encrypted_link_handshake, restore_encrypted_link_handshake_from_config,
+    EncryptedLinkHandshake, HandshakeProgress, DEFAULT_MAX_RECOVERY_ATTEMPTS,
+};
+pub use link::{
+    close_encrypted_link, restore_encrypted_link, restore_encrypted_link_from_config,
+    EncryptedLink, DEFAULT_MAX_SEND_RETRIES,
+};
+pub use snapshot::{EncryptedLinkHandshakeSnapshot, EncryptedLinkSnapshot};
