+++ paykit-lib/src/encrypted_link/handshake.rs
@@ -1,0 +1,549 @@
+use std::collections::VecDeque;
+
+use tracing::{debug, instrument, warn};
+
+use crate::{PaykitError, PublicKey, Result};
+
+use super::{
+    link::{EncryptedLink, DEFAULT_MAX_SEND_RETRIES},
+    paths::compute_private_payment_paths,
+    snapshot::EncryptedLinkHandshakeSnapshot,
+};
+
+/// Default maximum number of consecutive automatic recovery attempts before
+/// [`advance_handshake`] gives up and returns an error.
+///
+/// Override per-handshake via [`EncryptedLinkHandshake::set_max_recovery_attempts`].
+pub const DEFAULT_MAX_RECOVERY_ATTEMPTS: u32 = 3;
+
+/// Handle to an in-progress Noise handshake.
+///
+/// Created by [`initiate_encrypted_link`] (initiator) or
+/// [`accept_encrypted_link`] (responder). Drive the handshake forward by
+/// repeatedly calling [`advance_handshake`] until it returns
+/// [`HandshakeProgress::Complete`].
+///
+/// The caller controls the polling strategy — timing between retries, timeouts,
+/// back-off, etc. are all the caller's responsibility.
+///
+/// # Automatic recovery
+///
+/// If a homeserver write fails during the handshake (corrupting the internal
+/// Noise state), [`advance_handshake`] automatically restores from a
+/// pre-mutation snapshot and returns [`HandshakeProgress::Pending`] so the
+/// caller's polling loop retries transparently. The maximum number of
+/// consecutive recovery attempts is configurable via
+/// [`set_max_recovery_attempts`](Self::set_max_recovery_attempts) (default:
+/// [`DEFAULT_MAX_RECOVERY_ATTEMPTS`]).
+///
+/// # Session resumption
+///
+/// An in-progress handshake can be snapshotted via [`snapshot`](Self::snapshot)
+/// (or serialized directly via [`serialize`](Self::serialize)) and later
+/// restored with [`restore_encrypted_link_handshake`] or
+/// [`restore_encrypted_link_handshake_from_config`].
+///
+/// Restored handshakes always reset recovery tuning to defaults:
+/// `recovery_attempts` starts at `0` and `max_recovery_attempts` is set to
+/// [`DEFAULT_MAX_RECOVERY_ATTEMPTS`].
+pub struct EncryptedLinkHandshake {
+    /// The Noise session manager in handshake mode.
+    pub(crate) encryptor: pubky_noise::PubkyNoiseEncryptor,
+    /// The counterparty's public key (used for homeserver path construction).
+    pub(crate) remote_pubkey: PublicKey,
+    /// Shared Noise configuration needed for snapshot-based recovery.
+    pub(crate) config: std::sync::Arc<pubky_noise::PubkyNoiseConfig>,
+    /// Number of consecutive recovery attempts so far.
+    pub(crate) recovery_attempts: u32,
+    /// Maximum consecutive recovery attempts before giving up.
+    pub(crate) max_recovery_attempts: u32,
+}
+
+impl EncryptedLinkHandshake {
+    /// Set the maximum number of consecutive automatic recovery attempts
+    /// before [`advance_handshake`] gives up and returns
+    /// [`PaykitError::Transport`].
+    ///
+    /// The recovery-attempt counter resets to zero after every successful
+    /// handshake step.
+    /// Default: [`DEFAULT_MAX_RECOVERY_ATTEMPTS`] (3).
+    pub fn set_max_recovery_attempts(&mut self, max: u32) -> &mut Self {
+        self.max_recovery_attempts = max;
+        self
+    }
+
+    /// Capture the current handshake state as a serializable snapshot.
+    ///
+    /// The snapshot contains everything needed to restore and continue the
+    /// handshake later via [`restore_encrypted_link_handshake`] or
+    /// [`restore_encrypted_link_handshake_from_config`].
+    pub fn snapshot(&self) -> EncryptedLinkHandshakeSnapshot {
+        EncryptedLinkHandshakeSnapshot {
+            state: self.encryptor.snapshot(),
+            recipient: self.remote_pubkey.clone(),
+        }
+    }
+
+    /// Serialize the current handshake state to bytes for persistence.
+    ///
+    /// Convenience method equivalent to `self.snapshot().serialize()`.
+    pub fn serialize(&self) -> Vec<u8> {
+        self.snapshot().serialize()
+    }
+
+    /// Access the shared Noise configuration for this handshake.
+    ///
+    /// Useful for passing to [`restore_encrypted_link_handshake_from_config`]
+    /// when performing in-process recovery without an app restart.
+    pub fn config(&self) -> &std::sync::Arc<pubky_noise::PubkyNoiseConfig> {
+        &self.config
+    }
+}
+
+/// Result of a single [`advance_handshake`] step.
+pub enum HandshakeProgress {
+    /// Handshake is still in progress. The counterparty may not have written
+    /// their next message yet. Pass the returned handle back to
+    /// [`advance_handshake`] after a caller-chosen delay.
+    Pending(EncryptedLinkHandshake),
+
+    /// Handshake completed successfully. The [`EncryptedLink`] is ready for use
+    /// with [`crate::set_private_payment_envelope`] and [`crate::get_private_payment_envelope`].
+    Complete(EncryptedLink),
+}
+
+/// Initiates a Noise XX Encrypted Link Handshake with a counterparty
+/// (initiator role).
+///
+/// Initializes the encryption stack and creates a handshake context. The actual
+/// handshake messages are exchanged by repeatedly calling [`advance_handshake`]
+/// until it returns [`HandshakeProgress::Complete`].
+///
+/// Ephemeral keys are managed internally by the Noise stack — callers only need
+/// to provide their static identity key and the counterparty public key.
+///
+/// # Parameters
+/// - `session` — authenticated Pubky session for writing handshake messages
+///   (consumed; caller should `.clone()` if needed elsewhere).
+/// - `sender_secret_key` — 32-byte Ed25519 secret key of the local party.
+/// - `receiver_pubkey` — public key of the counterparty.
+/// - `outbox_client` — HTTP client for reading from the remote homeserver
+///   (consumed; caller should `.clone()` if needed elsewhere).
+///
+/// # Errors
+/// Returns [`PaykitError::Transport`] if the encryption stack cannot be
+/// initialized or if the context creation fails.
+#[instrument(skip(session, sender_secret_key, outbox_client))]
+pub fn initiate_encrypted_link(
+    session: pubky::PubkySession,
+    sender_secret_key: [u8; 32],
+    receiver_pubkey: &PublicKey,
+    outbox_client: pubky::Pubky,
+) -> Result<EncryptedLinkHandshake> {
+    debug!("initializing Encrypted Link handshake (initiator)");
+
+    let (write_path, read_path) =
+        compute_private_payment_paths(&sender_secret_key, receiver_pubkey);
+
+    let config = pubky_noise::PubkyNoiseConfig::new_with_paths(
+        sender_secret_key,
+        0,
+        "XX",
+        session,
+        write_path,
+        read_path,
+        outbox_client,
+    )
+    .map_err(|err| PaykitError::Transport {
+        context: format!("failed to create encryptor config: {err:?}"),
+        source: anyhow::anyhow!("pubky-noise PubkyNoiseConfig::new failed: {err:?}"),
+    })?;
+
+    let encryptor = pubky_noise::PubkyNoiseEncryptor::new(
+        config.clone(),
+        sender_secret_key,
+        true,
+        receiver_pubkey.clone(),
+    )
+    .map_err(|err| PaykitError::Transport {
+        context: format!("failed to initialize encryptor: {err:?}"),
+        source: anyhow::anyhow!("pubky-noise PubkyNoiseEncryptor::new failed: {err:?}"),
+    })?;
+
+    debug!("handshake context initialized (initiator)");
+    Ok(EncryptedLinkHandshake {
+        encryptor,
+        remote_pubkey: receiver_pubkey.clone(),
+        config,
+        recovery_attempts: 0,
+        max_recovery_attempts: DEFAULT_MAX_RECOVERY_ATTEMPTS,
+    })
+}
+
+/// Accepts a Noise XX Encrypted Link Handshake from a counterparty
+/// (responder role).
+///
+/// Initializes the encryption stack and creates a handshake context for the
+/// responder side. The actual handshake messages are exchanged by repeatedly
+/// calling [`advance_handshake`] until it returns [`HandshakeProgress::Complete`].
+///
+/// # Parameters
+/// - `session` — authenticated Pubky session for writing handshake messages
+///   (consumed; caller should `.clone()` if needed elsewhere).
+/// - `receiver_secret_key` — 32-byte Ed25519 secret key of the local party.
+/// - `sender_pubkey` — public key of the counterparty (the initiator).
+/// - `outbox_client` — HTTP client for reading from the remote homeserver
+///   (consumed; caller should `.clone()` if needed elsewhere).
+///
+/// # Errors
+/// Returns [`PaykitError::Transport`] if the encryption stack cannot be
+/// initialized or if the context creation fails.
+#[instrument(skip(session, receiver_secret_key, outbox_client))]
+pub fn accept_encrypted_link(
+    session: pubky::PubkySession,
+    receiver_secret_key: [u8; 32],
+    sender_pubkey: &PublicKey,
+    outbox_client: pubky::Pubky,
+) -> Result<EncryptedLinkHandshake> {
+    debug!("initializing Encrypted Link handshake (responder)");
+
+    let (write_path, read_path) =
+        compute_private_payment_paths(&receiver_secret_key, sender_pubkey);
+
+    let config = pubky_noise::PubkyNoiseConfig::new_with_paths(
+        receiver_secret_key,
+        0,
+        "XX",
+        session,
+        write_path,
+        read_path,
+        outbox_client,
+    )
+    .map_err(|err| PaykitError::Transport {
+        context: format!("failed to create encryptor config: {err:?}"),
+        source: anyhow::anyhow!("pubky-noise PubkyNoiseConfig::new failed: {err:?}"),
+    })?;
+
+    let encryptor = pubky_noise::PubkyNoiseEncryptor::new(
+        config.clone(),
+        receiver_secret_key,
+        false,
+        sender_pubkey.clone(),
+    )
+    .map_err(|err| PaykitError::Transport {
+        context: format!("failed to initialize encryptor: {err:?}"),
+        source: anyhow::anyhow!("pubky-noise PubkyNoiseEncryptor::new failed: {err:?}"),
+    })?;
+
+    debug!("handshake context initialized (responder)");
+    Ok(EncryptedLinkHandshake {
+        encryptor,
+        remote_pubkey: sender_pubkey.clone(),
+        config,
+        recovery_attempts: 0,
+        max_recovery_attempts: DEFAULT_MAX_RECOVERY_ATTEMPTS,
+    })
+}
+
+/// Advances the handshake by one step.
+///
+/// This function is **polling-safe**: calling it when the counterparty has not
+/// written their next message yet returns [`HandshakeProgress::Pending`] without
+/// corrupting internal state. The caller can safely retry after a delay.
+///
+/// # Automatic recovery
+///
+/// If the homeserver write fails during a handshake step
+/// (`HomeserverWriteError`), the internal Noise state is irreversibly
+/// corrupted. This function automatically recovers by restoring from the
+/// pre-mutation snapshot captured at the start of the failed step and returns
+/// [`HandshakeProgress::Pending`] so the caller's polling loop retries
+/// transparently.
+///
+/// The maximum number of **consecutive** recovery attempts is configurable via
+/// [`EncryptedLinkHandshake::set_max_recovery_attempts`] (default:
+/// [`DEFAULT_MAX_RECOVERY_ATTEMPTS`]). The recovery-attempt counter resets to
+/// zero after every successful step. If the limit is exceeded, the function returns
+/// [`PaykitError::Transport`].
+///
+/// # Polling strategy
+///
+/// The caller controls the polling strategy. Common patterns:
+///
+/// **Fixed interval:**
+/// ```ignore
+/// loop {
+///     match advance_handshake(handshake).await? {
+///         HandshakeProgress::Pending(h) => {
+///             handshake = h;
+///             tokio::time::sleep(Duration::from_millis(100)).await;
+///         }
+///         HandshakeProgress::Complete(link) => break link,
+///     }
+/// }
+/// ```
+///
+/// **With timeout:**
+/// ```ignore
+/// let deadline = Instant::now() + Duration::from_secs(60);
+/// loop {
+///     if Instant::now() > deadline {
+///         return Err(/* timeout */);
+///     }
+///     match advance_handshake(handshake).await? {
+///         HandshakeProgress::Pending(h) => {
+///             handshake = h;
+///             tokio::time::sleep(Duration::from_millis(100)).await;
+///         }
+///         HandshakeProgress::Complete(link) => break link,
+///     }
+/// }
+/// ```
+///
+/// # Parameters
+/// - `handshake` — the in-progress handshake handle (consumed; returned inside
+///   [`HandshakeProgress::Pending`] if the handshake is not yet finished).
+///
+/// # Errors
+/// - Returns [`PaykitError::Transport`] if the handshake processing fails, if
+///   the context is in an invalid state, or if automatic recovery is exhausted.
+#[instrument(skip(handshake))]
+pub async fn advance_handshake(mut handshake: EncryptedLinkHandshake) -> Result<HandshakeProgress> {
+    // Check whether the handshake has already finished.
+    if handshake.encryptor.is_handshake_complete() {
+        return finish_handshake(handshake);
+    }
+
+    // Process the next handshake step.
+    match handshake.encryptor.handle_handshake().await {
+        Ok(pubky_noise::HandshakeResult::Pending) => {
+            debug!("handshake step pending (waiting for counterparty)");
+            handshake.recovery_attempts = 0;
+            Ok(HandshakeProgress::Pending(handshake))
+        }
+        Ok(pubky_noise::HandshakeResult::Terminal) => {
+            debug!("handshake terminal, transitioning to transport");
+            finish_handshake(handshake)
+        }
+        Err(pubky_noise::PubkyNoiseError::HomeserverWriteError) => {
+            handshake.recovery_attempts += 1;
+
+            if handshake.recovery_attempts > handshake.max_recovery_attempts {
+                return Err(PaykitError::Transport {
+                    context: format!(
+                        "handshake recovery exhausted after {} consecutive attempts",
+                        handshake.max_recovery_attempts,
+                    ),
+                    source: anyhow::anyhow!(
+                        "HomeserverWriteError persisted beyond recovery limit ({})",
+                        handshake.max_recovery_attempts,
+                    ),
+                });
+            }
+
+            warn!(
+                attempts = handshake.recovery_attempts,
+                max = handshake.max_recovery_attempts,
+                "handshake write failed, attempting automatic recovery from snapshot"
+            );
+
+            let snapshot = handshake
+                .encryptor
+                .last_good_snapshot()
+                .cloned()
+                .ok_or_else(|| PaykitError::Transport {
+                    context: "handshake recovery failed: missing last-good snapshot".into(),
+                    source: anyhow::anyhow!(
+                        "pubky-noise returned HomeserverWriteError but no recovery snapshot"
+                    ),
+                })?;
+
+            let restored = pubky_noise::PubkyNoiseEncryptor::restore(
+                handshake.config.clone(),
+                snapshot,
+                handshake.remote_pubkey.clone(),
+            )
+            .await
+            .map_err(|err| PaykitError::Transport {
+                context: format!("handshake recovery via restore() failed: {err:?}"),
+                source: anyhow::anyhow!("restore after HomeserverWriteError failed: {err:?}"),
+            })?;
+
+            debug!("handshake recovered successfully, returning Pending");
+            Ok(HandshakeProgress::Pending(EncryptedLinkHandshake {
+                encryptor: restored,
+                config: handshake.config,
+                remote_pubkey: handshake.remote_pubkey,
+                recovery_attempts: handshake.recovery_attempts,
+                max_recovery_attempts: handshake.max_recovery_attempts,
+            }))
+        }
+        Err(err) => Err(PaykitError::Transport {
+            context: format!("handshake step failed: {err:?}"),
+            source: anyhow::anyhow!("pubky-noise handle_handshake failed: {err:?}"),
+        }),
+    }
+}
+
+/// Transitions a completed handshake into an [`EncryptedLink`].
+fn finish_handshake(mut handshake: EncryptedLinkHandshake) -> Result<HandshakeProgress> {
+    let _link_id =
+        handshake
+            .encryptor
+            .transition_transport()
+            .map_err(|err| PaykitError::Transport {
+                context: format!("failed to transition to transport mode: {err:?}"),
+                source: anyhow::anyhow!("pubky-noise transition_transport failed: {err:?}"),
+            })?;
+
+    debug!("Encrypted Link established");
+    Ok(HandshakeProgress::Complete(EncryptedLink {
+        encryptor: handshake.encryptor,
+        recipient: handshake.remote_pubkey,
+        config: handshake.config,
+        max_send_retries: DEFAULT_MAX_SEND_RETRIES,
+        pending_private_messages: VecDeque::new(),
+    }))
+}
+
+/// Restores an [`EncryptedLinkHandshake`] from a previously saved snapshot.
+///
+/// Use this to resume an in-progress handshake after an app restart. A fresh
+/// [`pubky_noise::PubkyNoiseConfig`] is built from the supplied session and key
+/// material, then replay restore reconstructs the handshake state from the
+/// persisted snapshot and homeserver data.
+///
+/// # Parameters
+/// - `session` — authenticated Pubky session for writing handshake messages
+///   (a fresh session after app restart).
+/// - `secret_key` — 32-byte Ed25519 secret key of the local party (same key
+///   used in the original [`initiate_encrypted_link`] or
+///   [`accept_encrypted_link`] call).
+/// - `remote_pubkey` — public key of the counterparty.
+/// - `outbox_client` — HTTP client for reading from the remote homeserver.
+/// - `snapshot` — saved in-progress handshake snapshot (from
+///   [`EncryptedLinkHandshake::snapshot`] or
+///   [`EncryptedLinkHandshakeSnapshot::deserialize`]).
+///
+/// The `remote_pubkey` must match `snapshot.recipient()`. A mismatch indicates
+/// inconsistent caller input and is rejected.
+///
+/// # Restore behavior
+///
+/// Restored handshakes always reset recovery tuning to defaults:
+/// - `recovery_attempts = 0`
+/// - `max_recovery_attempts = DEFAULT_MAX_RECOVERY_ATTEMPTS`
+///
+/// # Errors
+/// Returns [`PaykitError::Transport`] if the Noise configuration cannot be
+/// created or if the underlying `restore()` fails. Returns
+/// [`PaykitError::Validation`] when `remote_pubkey` does not match the
+/// recipient embedded in `snapshot`, or when the snapshot is not in handshake
+/// phase.
+#[instrument(skip(session, secret_key, outbox_client, snapshot))]
+pub async fn restore_encrypted_link_handshake(
+    session: pubky::PubkySession,
+    secret_key: [u8; 32],
+    remote_pubkey: &PublicKey,
+    outbox_client: pubky::Pubky,
+    snapshot: EncryptedLinkHandshakeSnapshot,
+) -> Result<EncryptedLinkHandshake> {
+    debug!("restoring Encrypted Link handshake from snapshot (raw params)");
+
+    let (write_path, read_path) = compute_private_payment_paths(&secret_key, remote_pubkey);
+
+    let config = pubky_noise::PubkyNoiseConfig::new_with_paths(
+        secret_key,
+        0,
+        "XX",
+        session,
+        write_path,
+        read_path,
+        outbox_client,
+    )
+    .map_err(|err| PaykitError::Transport {
+        context: format!("failed to create encryptor config for handshake restore: {err:?}"),
+        source: anyhow::anyhow!("pubky-noise PubkyNoiseConfig::new failed: {err:?}"),
+    })?;
+
+    restore_encrypted_link_handshake_inner(config, remote_pubkey, snapshot).await
+}
+
+/// Restores an [`EncryptedLinkHandshake`] from a previously saved snapshot
+/// using an existing Noise configuration.
+///
+/// This is the in-process variant of [`restore_encrypted_link_handshake`] — use
+/// it when the original `Arc<PubkyNoiseConfig>` is still available.
+///
+/// # Parameters
+/// - `config` — shared Noise configuration matching the original handshake
+///   session.
+/// - `remote_pubkey` — public key of the counterparty.
+/// - `snapshot` — saved in-progress handshake snapshot.
+///
+/// # Restore behavior
+///
+/// Restored handshakes always reset recovery tuning to defaults:
+/// - `recovery_attempts = 0`
+/// - `max_recovery_attempts = DEFAULT_MAX_RECOVERY_ATTEMPTS`
+///
+/// # Errors
+/// Returns [`PaykitError::Transport`] if the underlying `restore()` fails.
+/// Returns [`PaykitError::Validation`] when `remote_pubkey` does not match the
+/// recipient embedded in `snapshot`, or when the snapshot is not in handshake
+/// phase.
+#[instrument(skip(config, snapshot))]
+pub async fn restore_encrypted_link_handshake_from_config(
+    config: std::sync::Arc<pubky_noise::PubkyNoiseConfig>,
+    remote_pubkey: &PublicKey,
+    snapshot: EncryptedLinkHandshakeSnapshot,
+) -> Result<EncryptedLinkHandshake> {
+    debug!("restoring Encrypted Link handshake from snapshot (existing config)");
+    restore_encrypted_link_handshake_inner(config, remote_pubkey, snapshot).await
+}
+
+/// Shared implementation for both handshake restore variants.
+async fn restore_encrypted_link_handshake_inner(
+    config: std::sync::Arc<pubky_noise::PubkyNoiseConfig>,
+    remote_pubkey: &PublicKey,
+    snapshot: EncryptedLinkHandshakeSnapshot,
+) -> Result<EncryptedLinkHandshake> {
+    if snapshot.recipient() != remote_pubkey {
+        return Err(PaykitError::Validation(format!(
+            "remote_pubkey does not match snapshot recipient (remote={}, snapshot={})",
+            remote_pubkey,
+            snapshot.recipient(),
+        )));
+    }
+
+    if !matches!(
+        snapshot.state.phase,
+        pubky_noise::snow_crypto::NoisePhase::HandShake
+    ) {
+        return Err(PaykitError::Validation(format!(
+            "handshake restore requires handshake-phase snapshot, got {:?}",
+            snapshot.state.phase,
+        )));
+    }
+
+    let encryptor = pubky_noise::PubkyNoiseEncryptor::restore(
+        config.clone(),
+        snapshot.state,
+        remote_pubkey.clone(),
+    )
+    .await
+    .map_err(|err| PaykitError::Transport {
+        context: format!("failed to restore Encrypted Link handshake: {err:?}"),
+        source: anyhow::anyhow!("pubky-noise handshake restore failed: {err:?}"),
+    })?;
+
+    debug!("Encrypted Link handshake restored successfully (recovery tuning reset to defaults)");
+
+    Ok(EncryptedLinkHandshake {
+        encryptor,
+        remote_pubkey: remote_pubkey.clone(),
+        config,
+        recovery_attempts: 0,
+        max_recovery_attempts: DEFAULT_MAX_RECOVERY_ATTEMPTS,
+    })
+}
