+++ paykit-lib/src/tests/mod.rs
@@ -1,0 +1,257 @@
+use std::collections::HashMap;
+
+use crate::*;
+use pubky::PubkySession;
+use pubky_testnet::{embedded_postgres::EmbeddedPostgres, pubky::Keypair, EphemeralTestnet};
+use tokio::sync::{Mutex as TokioMutex, OnceCell};
+
+mod encrypted_link;
+mod private_payment_envelopes;
+mod public_payment_endpoints;
+mod receipt_access;
+
+static SHARED_POSTGRES: OnceCell<EmbeddedPostgres> = OnceCell::const_new();
+static TESTNET_BUILD_LOCK: TokioMutex<()> = TokioMutex::const_new(());
+
+async fn shared_postgres() -> &'static EmbeddedPostgres {
+    SHARED_POSTGRES
+        .get_or_init(|| async {
+            EmbeddedPostgres::start()
+                .await
+                .expect("failed to start embedded postgres")
+        })
+        .await
+}
+
+async fn build_testnet() -> EphemeralTestnet {
+    let _guard = TESTNET_BUILD_LOCK.lock().await;
+
+    let builder = if std::env::var_os("TEST_PUBKY_CONNECTION_STRING").is_some() {
+        EphemeralTestnet::builder()
+    } else {
+        let postgres = shared_postgres()
+            .await
+            .connection_string()
+            .expect("embedded postgres connection string should be valid");
+        EphemeralTestnet::builder().postgres(postgres)
+    };
+
+    builder.build().await.unwrap()
+}
+
+struct TestSetup {
+    _testnet: EphemeralTestnet,
+    session: PubkySession,
+    public_storage: pubky::PublicStorage,
+    raw_session: PubkySession,
+    public_key: PublicKey,
+}
+
+impl TestSetup {
+    async fn new() -> Self {
+        let testnet = build_testnet().await;
+
+        let homeserver = testnet.homeserver_app();
+        let sdk = testnet.sdk().unwrap();
+
+        let pair = Keypair::random();
+        let signer = sdk.signer(pair.clone());
+        let session = signer.signup(&homeserver.public_key(), None).await.unwrap();
+
+        let public_storage = sdk.public_storage();
+
+        Self {
+            _testnet: testnet,
+            session: session.clone(),
+            public_storage,
+            raw_session: session,
+            public_key: pair.public_key(),
+        }
+    }
+}
+
+/// Test setup that creates two users and initializes handshake handles
+/// without driving them to completion.
+struct InProgressHandshakeSetup {
+    _testnet: EphemeralTestnet,
+    initiator_session: PubkySession,
+    responder_session: PubkySession,
+    initiator_handshake: EncryptedLinkHandshake,
+    responder_handshake: EncryptedLinkHandshake,
+}
+
+impl InProgressHandshakeSetup {
+    async fn new() -> Self {
+        let testnet = build_testnet().await;
+        let homeserver = testnet.homeserver_app();
+
+        let initiator_sdk = testnet.sdk().unwrap();
+        let responder_sdk = testnet.sdk().unwrap();
+
+        let initiator_keypair = Keypair::random();
+        let initiator_signer = initiator_sdk.signer(initiator_keypair.clone());
+        let initiator_session = initiator_signer
+            .signup(&homeserver.public_key(), None)
+            .await
+            .unwrap();
+
+        let responder_keypair = Keypair::random();
+        let responder_signer = responder_sdk.signer(responder_keypair.clone());
+        let responder_session = responder_signer
+            .signup(&homeserver.public_key(), None)
+            .await
+            .unwrap();
+
+        let initiator_public_key = initiator_session.info().public_key();
+        let responder_public_key = responder_session.info().public_key();
+
+        let initiator_handshake = initiate_encrypted_link(
+            initiator_session.clone(),
+            initiator_keypair.secret_key(),
+            responder_public_key,
+            initiator_sdk,
+        )
+        .unwrap();
+
+        let responder_handshake = accept_encrypted_link(
+            responder_session.clone(),
+            responder_keypair.secret_key(),
+            initiator_public_key,
+            responder_sdk,
+        )
+        .unwrap();
+
+        Self {
+            _testnet: testnet,
+            initiator_session,
+            responder_session,
+            initiator_handshake,
+            responder_handshake,
+        }
+    }
+}
+
+/// Test setup for private (encrypted) payment flows.
+///
+/// Creates two users on the same ephemeral testnet, performs a full Noise XX
+/// handshake between them using the public `initiate_encrypted_link` /
+/// `accept_encrypted_link` / `advance_handshake` API, and produces ready-to-use
+/// [`EncryptedLink`] handles so that `set_private_payment_envelope` /
+/// `get_private_payment_envelope` can be exercised.
+struct PrivateTestSetup {
+    _testnet: EphemeralTestnet,
+    /// Sender's Encrypted Link (writes Private Payment Envelopes).
+    sender_link: EncryptedLink,
+    /// Sender's session (kept for cleanup via `signout`).
+    sender_session: PubkySession,
+    /// Receiver's Encrypted Link (reads Private Payment Envelopes).
+    receiver_link: EncryptedLink,
+    /// Receiver's session (kept for cleanup via `signout`).
+    receiver_session: PubkySession,
+}
+
+/// Drives a handshake to completion by polling `advance_handshake` with a
+/// short sleep between retries. Panics on timeout (10 s).
+async fn drive_handshake_to_completion(mut handshake: EncryptedLinkHandshake) -> EncryptedLink {
+    use std::time::{Duration, Instant};
+
+    let start = Instant::now();
+    let timeout = Duration::from_secs(10);
+
+    loop {
+        assert!(
+            start.elapsed() < timeout,
+            "handshake timed out after {timeout:?}"
+        );
+
+        match advance_handshake(handshake).await.unwrap() {
+            HandshakeProgress::Pending(h) => {
+                handshake = h;
+                tokio::time::sleep(Duration::from_millis(50)).await;
+            }
+            HandshakeProgress::Complete(link) => return link,
+        }
+    }
+}
+
+impl PrivateTestSetup {
+    async fn new() -> Self {
+        let testnet = build_testnet().await;
+        let homeserver = testnet.homeserver_app();
+
+        // Each user gets its own Pubky SDK instance.
+        let sender_sdk = testnet.sdk().unwrap();
+        let receiver_sdk = testnet.sdk().unwrap();
+
+        // Sign up two independent users.
+        let sender_keypair = Keypair::random();
+        let sender_signer = sender_sdk.signer(sender_keypair.clone());
+        let sender_session = sender_signer
+            .signup(&homeserver.public_key(), None)
+            .await
+            .unwrap();
+
+        let receiver_keypair = Keypair::random();
+        let receiver_signer = receiver_sdk.signer(receiver_keypair.clone());
+        let receiver_session = receiver_signer
+            .signup(&homeserver.public_key(), None)
+            .await
+            .unwrap();
+
+        let sender_public_key = sender_session.info().public_key();
+        let receiver_public_key = receiver_session.info().public_key();
+
+        // Initiate handshake from sender side.
+        let sender_handshake = initiate_encrypted_link(
+            sender_session.clone(),
+            sender_keypair.secret_key(),
+            receiver_public_key,
+            sender_sdk,
+        )
+        .unwrap();
+
+        // Accept handshake from receiver side.
+        let receiver_handshake = accept_encrypted_link(
+            receiver_session.clone(),
+            receiver_keypair.secret_key(),
+            sender_public_key,
+            receiver_sdk,
+        )
+        .unwrap();
+
+        // Drive both handshakes to completion concurrently.
+        let (sender_link, receiver_link) = tokio::join!(
+            drive_handshake_to_completion(sender_handshake),
+            drive_handshake_to_completion(receiver_handshake),
+        );
+
+        Self {
+            _testnet: testnet,
+            sender_link,
+            sender_session,
+            receiver_link,
+            receiver_session,
+        }
+    }
+}
+
+// ── Private Payment Envelopes tests ──────────────────────────────────────────
+
+fn private_payment_envelope(
+    payment_endpoints: &HashMap<PaymentEndpointIdentifier, PaymentEndpointPayload>,
+) -> PrivatePaymentEnvelope {
+    PrivatePaymentEnvelope::new(PaymentReference::new_v4(), payment_endpoints.clone())
+}
+
+const TEST_RECEIPT_ACCESS_JSON: &str = r#"{"version":1,"kind":"paykit.receipt_access","reference":"550e8400-e29b-41d4-a716-446655440000"}"#;
+
+async fn send_raw_private_message(link: &mut EncryptedLink, json: &str) {
+    assert!(
+        json.len() <= pubky_noise::snow_crypto::PUBKY_NOISE_MSG_LEN,
+        "test raw message exceeds pubky-noise message size"
+    );
+    link.encryptor
+        .send_message(json.as_bytes())
+        .await
+        .expect("raw private message should send");
+}
